re-arrange the files
