+++ src/rvesimulator/__init__.py
@@ -1,4 +1,4 @@
-"""rvesimulator - a python package for simulating Representative Volume Elements
+"""rvesimulator - python package for simulating Representative Volume Elements
 (RVEs) via commercial finite element software Abaqus.
 
 This package provides tools for generating micro-structures, simulating RVEs,
@@ -31,5 +31,4 @@ from .benchmarks.cddm_rve import CDDM_RVE
 from .benchmarks.hollow_plate_sve import (ElasticRegularLoads,
                                           VonMisesPlasticPathLoads,
                                           VonMisesPlasticRegularLoads)
-from .benchmarks.pppe_no_cohesive_rve import PPPEMixtureNoCohesive
 from .microstructure import circle_particles, sphere_particles
